Base authentication files.
Still need to add routes and process for filtering user authentications.
Also need to edit the User class so that a unique, anonymous ID is used so
that people's emails aren't exposed through the browser. (ie:
/user/1234141 instead of /user/email@email.com).
Also need to add encryption for User's passwords via UserController and
add registration paths via the UserController.